closes issue https://github.com/redixhumayun/simpledb/issues/1
This commit revamps the iterator mechanism so that it works correctly.
Earlier, the iterator would be initialized to slot 0 but this was a problem because calling next() on that iterator would move it to slot 1.
The Java implementation set it to -1 but using Option<usize> in Rust is nicer so now its being set to None initially.
When next() is called, this Option is matched over and the correct value set.