Fixes Txn Multi-Threading Test Failure
This commit fixes the txn multi-threading failure that was plaguing CI after removing the serializable lock around BufferManager

Now, there is a global mutex within BufferManager temporarily to ensure correctness for that test. The test that was failing was `test_transaction_multi_threaded_single_reader_single_writer`.

This commit also adds docs and an issue around how to get rid of the serializable lock.